add Type2da.TYPE_NONE
It's not really being used though. Instead, the menu-items are disabled if
a 2da has not been loaded yet. It was going to be used, however, for a
right-click on the tree (to disable the context menu items if a 2da is not
loaded) but I've discarded the idea of turning this into a 2da row-editor,
although the code to start doing that is present here but commented out.